Reuse util functions
