+++ tests/subscriptions/test_builder_mode_state.py
@@ -1,35 +1,12 @@
 from datetime import timedelta
 from typing import Sequence, Tuple
-from uuid import UUID
 
 import pytest
 
 from snuba import settings, state
-from snuba.subscriptions.data import (
-    PartitionId,
-    SnQLSubscriptionData,
-    Subscription,
-    SubscriptionIdentifier,
-)
+from snuba.subscriptions.data import Subscription
 from snuba.subscriptions.scheduler import TaskBuilderMode, TaskBuilderModeState
-
-UUIDS = [
-    UUID("fac82541-049f-4435-982d-819082761a53"),
-    UUID("49215ec6-939e-41e9-a209-f09b5514e884"),
-]
-
-
-def build_subscription(resolution: timedelta, sequence: int) -> Subscription:
-    return Subscription(
-        SubscriptionIdentifier(PartitionId(1), UUIDS[sequence]),
-        SnQLSubscriptionData(
-            project_id=1,
-            time_window=timedelta(minutes=5),
-            resolution=resolution,
-            query="MATCH events SELECT count()",
-        ),
-    )
-
+from tests.subscriptions.subscriptions_utils import build_subscription
 
 ALIGNED_TIMESTAMP = 1625527800  # Aligned to start of a minute
