+++ tests/subscriptions/test_task_builder.py
@@ -1,16 +1,10 @@
 from datetime import datetime, timedelta
 from typing import Sequence, Tuple
-from uuid import UUID
 
 import pytest
 
 from snuba import state
-from snuba.subscriptions.data import (
-    PartitionId,
-    SnQLSubscriptionData,
-    Subscription,
-    SubscriptionIdentifier,
-)
+from snuba.subscriptions.data import Subscription
 from snuba.subscriptions.scheduler import (
     DelegateTaskBuilder,
     ImmediateTaskBuilder,
@@ -19,24 +13,7 @@ from snuba.subscriptions.scheduler import (
     TaskBuilder,
 )
 from snuba.utils.scheduler import ScheduledTask
-
-UUIDS = [
-    UUID("fac82541-049f-4435-982d-819082761a53"),
-    UUID("49215ec6-939e-41e9-a209-f09b5514e884"),
-]
-
-
-def build_subscription(resolution: timedelta, sequence: int) -> Subscription:
-    return Subscription(
-        SubscriptionIdentifier(PartitionId(1), UUIDS[sequence]),
-        SnQLSubscriptionData(
-            project_id=1,
-            time_window=timedelta(minutes=5),
-            resolution=resolution,
-            query="MATCH events SELECT count()",
-        ),
-    )
-
+from tests.subscriptions.subscriptions_utils import UUIDS, build_subscription
 
 ALIGNED_TIMESTAMP = 1625518080  # Aligned to start of a minute
